requirements.txt
google-auth==2.38.0
gspread==6.2.0
python-dotenv==1.1.0
streamlit==1.44.1
